autoincrement :)

+++ src/utils.js
@@ -1,9 +1,19 @@
 const Path = require("path");
 const fs = require("fs");
+const ErrorDb = require('./structures/error');
 
 module.exports = {
+    exists: (path) => {
+        return fs.existsSync(path)
+    },
     create: (path, name) => {
         if (!fs.existsSync(Path.join(path, `${name}.json`))) return fs.writeFileSync(Path.join(path, `${name}.json`), "{}");
+    },
+    delete: (path, name) => {
+      if(!fs.existsSync(Path.join(path, `${name}.json`))) throw ErrorDb('Cannot delete a database that doesn\'t exist.');
+      if(name === "autoincrementhandler") return console.warn("Warning, you are deleting autoincrementhandler.json, this may result in unexcepted results when using autoincrement. If you really want to delete this database, please do it manually.");
+      fs.unlinkSync(Path.join(path, `${name}.json`));
+      return true; 
     },
     checkName: (name) => {
         if (typeof name !== "string") return false;
@@ -27,7 +37,16 @@ module.exports = {
         if (twoDots) return false;
         return true;
     },
+    deleteData: (id, path) => {
+        if(!fs.existsSync(path)) throw new ErrorDb('Cannot find the database file.');
+        let allData = JSON.parse(fs.readFileSync(path, 'UTF8'));
+        if(!allData[id]) throw new ErrorDb('Cannot delete a entry that does\'t exist.')
+        delete allData[id]
+        fs.writeFileSync(path, JSON.stringify(allData));
+
+    },
     createData: (id, path, sValue) => {
+        if(!fs.existsSync(path)) throw new ErrorDb('Cannot find the database file.');
         let allData = JSON.parse(fs.readFileSync(path, 'UTF8'));
         id = id.split(".");
         let info = "";
@@ -47,9 +66,13 @@ module.exports = {
         return allData;
     },
     readAllData: (path) => {
+        if(!fs.existsSync(path)) throw new ErrorDb('Cannot find the database file.');
+        
         return JSON.parse(fs.readFileSync(path, 'UTF8'));
     },
     getData: (id, path, all = false) => {
+        if(!fs.existsSync(path)) throw new ErrorDb('Cannot find the database file.');
+
         let allData = JSON.parse(fs.readFileSync(path, 'UTF8'));
         if (id === "*" && all) return allData;
         id = id.split(".");
@@ -63,6 +86,8 @@ module.exports = {
         return eval(`allData${info}`);
     },
     setData: (id, path, data) => {
+        if(!fs.existsSync(path)) throw new ErrorDb('Cannot find the database file.');
+
         let allData = JSON.parse(fs.readFileSync(path, 'UTF8'));
         id = id.split(".");
         let info = "";
@@ -80,6 +105,8 @@ module.exports = {
         return allData;
     },
     saveData: (path, data) => {
+        if(!fs.existsSync(path)) throw new ErrorDb('Cannot find the database file.');
+
         fs.writeFileSync(path, JSON.stringify(data));
     }
 }
